Creates plots comparing the length frequency distribution of model catch and biomass against data

Creates plots comparing the length frequency distribution of model catch and biomass against data

plot_length_freq(
  param.dir,
  run.dir,
  run.name,
  catch.ref,
  survey.ref,
  ref.years,
  bin.width = 1,
  out.dir,
  out.name,
  param.ls,
  speciesCodes = NULL,
  fleetCodes = NULL
)

Arguments

param.dir

string. Path to location of atlantis parameter files

run.dir

character vector of path to location of atlantis output files

run.name

character vector of run name corresponding to run.dir

catch.ref

dataframe that contains reference catch data with format (Code|len.lower|len.upper|cum.prob)

survey.ref

dataframe that contains reference biomass data with format (Code|len.lower|len.upper|cum.prob)

ref.years

numeric vector with first and last year (years from start) for comparison period

bin.width

numeric. Width of bins (in cms) in which to bin Fish length. Default = 1

out.dir

string. path to desired location of post-processed output

out.name

string. name for output file prefix

param.ls

list generated from get_atl_paramfiles()

speciesCodes

character vector. List of species to make plots for. Default = NULL (All species)

fleetCodes

character vector. List of fleets to make plots for., Default = NULL (All fleets)

Value

PDF of all fleet x species combination cumulative length frequency distributions

Author: Joseph Caracappa